chore: move prometheus http to http mod
