+++ src/servers/src/grpc/prom_query_gateway.rs
@@ -35,9 +35,8 @@ use tonic::{Request, Response};
 use crate::error::InvalidQuerySnafu;
 use crate::grpc::greptime_handler::{auth, create_query_context};
 use crate::grpc::TonicResult;
-use crate::prometheus::{
-    retrieve_metric_name_and_result_type, PrometheusHandlerRef, PrometheusJsonResponse,
-};
+use crate::http::prometheus::{retrieve_metric_name_and_result_type, PrometheusJsonResponse};
+use crate::prometheus_handler::PrometheusHandlerRef;
 
 pub struct PrometheusGatewayService {
     handler: PrometheusHandlerRef,
