+++ tests-integration/tests/http.rs
@@ -20,8 +20,8 @@ use axum_test_helper::TestClient;
 use common_error::status_code::StatusCode as ErrorCode;
 use serde_json::json;
 use servers::http::handler::HealthResponse;
+use servers::http::prometheus::{PrometheusJsonResponse, PrometheusResponse};
 use servers::http::{JsonOutput, JsonResponse};
-use servers::prometheus::{PrometheusJsonResponse, PrometheusResponse};
 use tests_integration::test_util::{
     setup_test_http_app, setup_test_http_app_with_frontend,
     setup_test_http_app_with_frontend_and_user_provider, setup_test_prom_app_with_frontend,
